move requests to XUIConfig
